Added clickable navigation to Favourite Trips and new Favorites menu item
Made the "Favourite Trips" heading on the home screen clickable so users
can tap it to navigate to the full favourites screen. Also added a new
"Favorites" item in the drawer menu for quick access to saved trips and
preferred routes. Updated the Heading component to accept a modifier
parameter to support click handlers.